Pushing

+++ src/Renderer/DirectX/main.cpp
@@ -1,16 +1,16 @@
-
+ 
 #include <windows.h>
 #include <thread>
-#include "graphics.h"
+#include "graphics.h"    
 
 #include "IUnityInterface.h"
 #include "IUnityGraphics.h" 
 
-#include <map>
+#include <map> 
 
 static HWND ghWnd = NULL;
 static HWND selectedWnd = NULL;
-
+  
 HINSTANCE hInstance;
 std::thread* myThread = nullptr; 
 static std::map<int, HWND> windowIds;
@@ -44,12 +44,12 @@ LRESULT CALLBACK WindowProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lPara
         // this message is read when the window is closed
         case WM_DESTROY:
 
-			break; 
+			break;  
     } 
 
     // Handle any messages the switch statement didn't
     return DefWindowProc (hWnd, message, wParam, lParam);
-}
+}  
 
 HWND CreateNewWindow(LPCWSTR titlestr, int width, int height, bool noStyle) {	
 	LPCSTR title = " ";
@@ -67,7 +67,7 @@ HWND CreateNewWindow(LPCWSTR titlestr, int width, int height, bool noStyle) {
 	windowClass.cbSize = sizeof(WNDCLASSEX);
 	windowClass.style = CS_HREDRAW | CS_VREDRAW | CS_OWNDC;  
 	windowClass.lpfnWndProc = (WNDPROC) WindowProc;  
-	windowClass.cbClsExtra = 0;  
+	windowClass.cbClsExtra = 0;   
 	windowClass.cbWndExtra = 0;  
 	windowClass.hInstance = hInstance;  
 	windowClass.hIcon = LoadIcon(NULL, IDI_WINLOGO);  
@@ -83,7 +83,7 @@ HWND CreateNewWindow(LPCWSTR titlestr, int width, int height, bool noStyle) {
 
 	if (!RegisterClassEx(&windowClass)) {  
 		return NULL; 
-	}
+	}  
 
 	RECT wr = {0, 0, width, height};
 	AdjustWindowRect(&wr, WS_OVERLAPPEDWINDOW, FALSE);
@@ -171,35 +171,35 @@ extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API StopWindowById(int wi
 		myThread = nullptr;
 	}
 }
-
+ 
 extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetWindowRectById(int windowId, int left, int top, int width, int height) {
 	if (windowIds.count(windowId)) {
 		HWND hwnd = windowIds[windowId];
 		windowGraphics[hwnd].SetSize(width, height);
 		SetWindowPos(hwnd, 0, left, top, width, height, 0);
-	} 
+	}   
 }
 extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SendTextureIdToPluginByIdLeft(int windowId, void* texturePtr) {
-	if (windowIds.count(windowId)) {
+	if (windowIds.count(windowId)) { 
 		selectedWnd = windowIds[windowId];
 		windowGraphics[selectedWnd].SetTexturePtrLeft(texturePtr);
-	}
+	}     
 }
-extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetAffineTransform(int windowID, double* affineTransform) {
-	if (windowIds.count(windowID)) {
+extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetAffineTransform(int windowID, float* affineTransform) {
+	if (windowIds.count(windowID)) { 
 		selectedWnd = windowIds[windowID];
 		windowGraphics[selectedWnd].SetAffine(affineTransform);
-	}
-}
+	} 
+} 
 extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetRequiredValuesById(int windowID,
 float leftUvToRectX[],// = { 0.0 };
-float leftUvToRectY[],// = { 0.0 };
-float rightUvToRectX[],// = { 0.0 };
+float leftUvToRectY[],// = { 0.0 }; 
+float rightUvToRectX[],// = { 0.0 };   
 float rightUvToRectY[],// = { 0.0 };
 float CameraMatrixLeft[],// = { 0.0 };
 float CameraMatrixRight[],// = { 0.0 };
 float leftOffset[],// = { 0.0 };
-float rightOffset[],// = { 0.0 };
+float rightOffset[],// = { 0.0 };  
 float eyeBorders[]) {
 	if (windowIds.count(windowID)) {
 		selectedWnd = windowIds[windowID];
@@ -210,8 +210,8 @@ extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SendTextureIdToPlugin
 	if (windowIds.count(windowId)) {
 		selectedWnd = windowIds[windowId];
 		windowGraphics[selectedWnd].SetTexturePtrRight(texturePtr);
-	}
-}
+	}  
+}   
 static void SetAttributes(HWND hwnd, int colorKey, byte alpha, int flags) {
 	if (flags == 0) {
 		SetWindowLong(hwnd, GWL_EXSTYLE, GetWindowLong(hwnd, GWL_EXSTYLE) & ~WS_EX_LAYERED);
@@ -221,7 +221,7 @@ static void SetAttributes(HWND hwnd, int colorKey, byte alpha, int flags) {
 		if (flags == 1) {
 			SetLayeredWindowAttributes(hwnd, coloref, alpha, LWA_COLORKEY);
 		} else if (flags == 2) {
-			SetLayeredWindowAttributes(hwnd, coloref, alpha, LWA_ALPHA);
+			SetLayeredWindowAttributes(hwnd, coloref, alpha, LWA_ALPHA); 
 		} else if (flags == 3) {
 			SetLayeredWindowAttributes(hwnd, coloref, alpha, LWA_COLORKEY | LWA_ALPHA);
 		}
@@ -238,7 +238,7 @@ extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetWindowAttributesBy
 	if (windowIds.count(windowId)) {
 		SetAttributes(windowIds[windowId], colorKey, alpha, flags);
 	}
-} 
+}  
 
 extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API SetColorFormat(int colorFormat) {
 	if (colorFormat == 0) {
